Fix some edge cases of usage of static variables
