+++ src/main/java/com/team766/web/dashboard/Widget.java
@@ -8,6 +8,7 @@ import java.util.stream.Collectors;
 public abstract class Widget {
 	public static final int DEFAULT_SORT_ORDER = 0;
 
+	private static int c_orderCounter = 0;
 	private static Map<Widget, Long> c_widgets =
 		Collections.synchronizedMap(new WeakHashMap<Widget, Long>());
 
@@ -21,7 +22,9 @@ public abstract class Widget {
 	}
 
 	public Widget(int sortOrder) {
-		c_widgets.put(this, (((long)sortOrder) << 32) | c_widgets.size());
+		synchronized (c_widgets) {
+			c_widgets.put(this, (((long)sortOrder) << 32) | (c_orderCounter++));
+		}
 	}
 
 	public abstract String render();
